Crate fragnova_runtime

The Runtime of the Fragnova Node.

The runtime for a Substrate node contains all of the business logic for executing transactions, saving state transitions, and interacting with the outer node.

Footnotes

Weights

In Substrate, weight is a unit of measurement to measure computation time.

“Built into FRAME, it is hardcoded that 10**12 weight is equivalent to 1 second of computation time [on the reference hardware]” - https://www.youtube.com/watch?v=i3zW4wGexAc&t=138s

Therefore, one unit of weight is one picosecond of computation time on the reference hardware. Source: https://docs.substrate.io/reference/glossary/#weight

What does “Maximum Block Weight” mean?

The maximum block weight is the maximum amount of computation time that a Node is allowed to spend in constructing a block.

In Substrate, the maximum block weight should be equivalent to one-third of the target block time with an allocation of:

One third for block construction

One third for network propagation

One third for import and verification

Source: https://docs.substrate.io/reference/glossary/#weight

Transaction lifecycle (i.e the sequence of events that happens when a transaction/extrinsic is sent to a Substrate Blockchain):

The transaction (i.e the payload) is sent to a single Substrate Node using an RPC

The transaction enters the transaction pool of this Substrate Node 2a. In the transaction pool, the payload is checked to be valid in validate_transaction()

The Substrate Node gossips this transaction to peer nodes using libp2p 3a. The peer nodes also check if this transaction is valid by running validate_transaction()

One of the nodes creates a block which will include the transaction.

This “block author node” will gossip the block to peer nodes using libp2p

The gossiped block will be imported by the peer nodes (and surprisingly by the “block author node” also). 6a. All the transactions of the imported block will be executed. (This includes the transaction-in-question obviously)

Source: https://www.youtube.com/watch?v=3pfM0GOp02c&ab_channel=ParityTech

TODO Review - Notice that once the block is imported (i.e step 6a) - all its transactions/extrinsics get executed right-away (if we take what the presenter said in the YouTube video completely literally). It does not validate the transactions (i.e perform validate_transaction on them) in the block before executing them!

Transaction Format

Please see the types UncheckedExtrinsic and SignedExtra below to know what the transaction format should be when submitting a transaction to a Node of this Blockchain via RPC

Modules

api

opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core data structures.

Macros

construct_runtime

Construct a runtime, with the given name and the given pallets.

parameter_types

Create new implementations of the Get trait.

Structs

ApprovalDeposit

The amount of funds that must be reserved when creating a new approval.

AssetDeposit

The basic amount of funds that must be reserved for an asset.

BaseCallFilter

BlockExecutionWeight

Time to execute an empty block. Calculated by multiplying the Average with 1.0 and adding 0.

BlockHashCount

Maximum number of block number to block hash mappings to keep (oldest pruned first).

BlockLength

max_with_normal_ratio(max: u32, normal: Perbill) creates a new BlockLength with max for Operational & Mandatory and normal * max for Normal.

BlockWeights

Set the “target block weight” for the Fragnova Blockchain.

DeletionQueueDepth

The maximum number of contracts that can be pending for deletion.

DeletionWeightLimit

The maximum amount of weight that can be consumed per block for lazy trie removal.

DepositPerByte

The amount of balance a caller (here “caller” refers to a “smart-contract account”) has to pay for each byte of storage.

DepositPerItem

The amount of balance a caller (here “caller” refers to a “smart-contract account”) has to pay for each storage item.

ExistentialDeposit

The minimum amount required to keep an account open.

ExtrinsicBaseWeight

Time to execute a NO-OP extrinsic, for example System::remark. Calculated by multiplying the Average with 1.0 and adding 0.

GenesisConfig

IdentityFee

Implementor of WeightToFee that maps one unit of weight to one unit of fee.

IndexDeposit

TODO: Documentation

IsTransferable

Whether an account can voluntarily transfer any of its balance to another account

MaxAdditionalFields

Maximum number of additional fields that may be stored in an ID. Needed to bound the I/O required to access an identity, but can be pretty high.

MaxAuthorities

The maximum number of authorities that pallet_aura can hold.

MaxLocks

The maximum number of locks that should exist on an account. Not strictly enforced, but used for weight estimation.

MaxRegistrars

Maxmimum number of registrars allowed in the system. Needed to bound the complexity of, e.g., updating judgements.

MetadataDepositBase

The basic amount of funds that must be reserved when adding metadata to your asset.

MetadataDepositPerByte

The additional funds that must be reserved for the number of bytes you store in your asset’s metadata.

MinimumPeriod

TODO: Documentation

OperationalFeeMultiplier

A fee mulitplier for Operational extrinsics to compute “virtual tip” to boost their priority

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Perbill

A fixed point representation of a number in the range [0, 1].

Permill

A fixed point representation of a number in the range [0, 1].

RocksDbWeight

By default, Substrate uses RocksDB, so this will be the weight used throughout the runtime.

RootNamespace

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

This is used as an identifier of the chain. 42 is the generic substrate prefix.

Schedule

Cost schedule and limits.

StorageInfo

Metadata about storage from the runtime.

StringLimit

The maximum length of a name or symbol of an asset stored on-chain.

Version

Version of the runtime.

Weight

Enums

BalancesCall

Contains one variant per dispatchable that can be called by an extrinsic.

FragmentsCall

Contains one variant per dispatchable that can be called by an extrinsic.

OriginCaller

ProtosCall

Contains one variant per dispatchable that can be called by an extrinsic.

RuntimeCall

RuntimeEvent

TimestampCall

Contains one variant per dispatchable that can be called by an extrinsic.

Constants

CENTS

TODO: Documentation

CONTRACTS_DEBUG_OUTPUT

Prints debug output of the contracts pallet to stdout if the node is started with -lruntime::contracts=debug.

DAYS

Number of blocks that would be added to the Blockchain on average, when one day passes

DOLLARS

TODO: Documentation

HOURS

Number of blocks that would be added to the Blockchain on average, when one hour passes

MAXIMUM_BLOCK_LENGTH

The maximum possible length (in bytes) that a Fragnova Block can be

MAXIMUM_METADATA_DATA_LENGTH

Maximum length (in bytes) that the metadata data of a Proto-Fragment / Fragment Definition / Fragment Instance can be

MAXIMUM_NESTED_CALL_DEPTH_LEVEL

Maximum permitted depth level that a descendant call of the outermost call of an extrinsic can be

MILLICENTS

TODO: Documentation

MILLISECS_PER_BLOCK

This determines the average expected block time that we are targeting. Blocks will be produced at a minimum duration defined by SLOT_DURATION. SLOT_DURATION is picked up by pallet_timestamp which is in turn picked up by pallet_aura to implement fn slot_duration().

MINUTES

Number of blocks that would be added to the Blockchain on average, when one minute passes

SLOT_DURATION

Blocks will be produced at a minimum duration defined by SLOT_DURATION.

VERSION

To learn more about runtime versioning and what each of the following value means: https://docs.substrate.io/v3/runtime/upgrades#runtime-versioning

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage

Complex storage builder stuff.

Contains

A trait for querying whether a type can be said to “contain” a value.

KeyOwnerProofSystem

Something which can compute and check proofs of a historical key owner and return full identification data of that key owner.

Randomness

A trait that is able to provide randomness.

StorageValue

A trait for working with macro-generated storage values under the substrate storage API.

Functions

deposit

The amount of balance a caller has to pay for calling an extrinsic with bytes bytes and .

native_version

The version information is used to identify this runtime when compiled natively.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountIndex

Related to Index pallet

Accounts

AccountsConfig

Address

The address format for describing accounts.

Aliases

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Assets

AssetsConfig

Aura

AuraConfig

Balance

Balance of an account.

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockNumber

An index to a block.

Clusters

Contracts

Detach

DetachConfig

Executive

Executive: handles dispatch to the various modules.

Fragments

Grandpa

GrandpaConfig

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

Identity

Index

Index of a transaction in the chain.

Indices

IndicesConfig

Multisig

Oracle

OracleConfig

Protos

Proxy

RandomnessCollectiveFlip

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

The payload being signed in transactions.

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

UncheckedExtrinsic

The type (i.e “format”) that an unchecked extrinsic must have.

Utility