Enum fragnova_runtime::ProtosCall

pub enum ProtosCall<T>where
    T: Config,
    <T as Config>::AccountId: UncheckedFrom<<T as Config>::Hash> + AsRef<[u8]>,{
    upload {
        references: Vec<[u8; 32], Global>,
        category: Categories,
        tags: BoundedVec<BoundedVec<u8, <T as Config>::StringLimit>, <T as Config>::MaxTags>,
        linked_asset: Option<LinkedAsset>,
        license: UsageLicense<<T as Config>::AccountId>,
        cluster: Option<[u8; 16]>,
        data: ProtoData,
    },
    patch {
        proto_hash: [u8; 32],
        license: Option<UsageLicense<<T as Config>::AccountId>>,
        new_references: Vec<[u8; 32], Global>,
        tags: Option<BoundedVec<BoundedVec<u8, <T as Config>::StringLimit>, <T as Config>::MaxTags>>,
        data: Option<ProtoData>,
    },
    transfer {
        proto_hash: [u8; 32],
        new_owner: <T as Config>::AccountId,
    },
    set_metadata {
        proto_hash: [u8; 32],
        metadata_key: BoundedVec<u8, <T as Config>::StringLimit>,
        data: Vec<u8, Global>,
    },
    detach {
        proto_hashes: Vec<[u8; 32], Global>,
        target_chain: SupportedChains,
        target_account: BoundedVec<u8, <T as Config>::DetachAccountLimit>,
    },
    ban {
        proto_hash: [u8; 32],
    },
    // some variants omitted
}

Contains one variant per dispatchable that can be called by an extrinsic.

Variants

upload

Fields

references: Vec<[u8; 32], Global>

category: Categories

tags: BoundedVec<BoundedVec<u8, <T as Config>::StringLimit>, <T as Config>::MaxTags>

linked_asset: Option<LinkedAsset>

license: UsageLicense<<T as Config>::AccountId>

cluster: Option<[u8; 16]>

data: ProtoData

Upload a Proto-Fragment onto the Blockchain. Furthermore, this function also indexes data in the Blockchain’s Database and makes it available via bitswap (IPFS) directly from every chain node permanently.

Arguments

origin - The origin of the extrinsic function

references - List of other Proto-Fragments used to create the Proto-Fragment

categories - Category type of the Proto-Fragment

tags - List of tags to tag the Proto-Fragment with

linked_asset (optional) - An asset that is linked with the Proto-Fragment (e.g an ERC-721 Contract)

license - Enum indicating how the Proto-Fragment can be used. NOTE: If None, the Proto-Fragment can’t be included into other Proto-Fragments

cluster - the Cluster id the proto belongs to (Optional)

data - Data of the Proto-Fragment

patch

Fields

proto_hash: [u8; 32]

license: Option<UsageLicense<<T as Config>::AccountId>>

new_references: Vec<[u8; 32], Global>

tags: Option<BoundedVec<BoundedVec<u8, <T as Config>::StringLimit>, <T as Config>::MaxTags>>

data: Option<ProtoData>

Patch an existing Proto-Fragment (by appending the hash of data to the Vector field patches of the existing Proto-Fragment’s Struct Instance) Furthermore, this function also indexes data in the Blockchain’s Database and stores it in the IPFS

Arguments

origin - The origin of the extrinsic function

proto_hash - Existing Proto-Fragment’s hash

license (optional) - If this value is not None, the existing Proto-Fragment’s current license is overwritten to this value

new_references - List of New Proto-Fragments that was used to create the patch (data needs to be populated, or it has no effect)

tags (optional) - List of tags to overwrite the Proto-Fragment’s current list of tags with, if not None.

data - Data of the Proto-Fragment

transfer

Fields

proto_hash: [u8; 32]

new_owner: <T as Config>::AccountId

Transfer the ownership of a Proto-Fragment from origin to new_owner

Arguments

origin - The origin of the extrinsic function

proto_hash - The hash of the data of the Proto-Fragment to transfer

new_owner - The Account ID to transfer the Proto-Fragment to

set_metadata

Fields

proto_hash: [u8; 32]

metadata_key: BoundedVec<u8, <T as Config>::StringLimit>

data: Vec<u8, Global>

Alters the metadata of a Proto-Fragment (whose hash is proto_hash) by adding or modifying a key-value pair (metadata_key.clone,blake2_256(&data.encode())) to the BTreeMap field metadata of the existing Proto-Fragment’s Struct Instance. Furthermore, this function also indexes data in the Blockchain’s Database and stores it in the IPFS To successfully patch a Proto-Fragment, the auth provided must be valid. Otherwise, an error is returned

Arguments

origin - The origin of the extrinsic / dispatchable function

proto_hash - The hash of the Proto-Fragment

metadata_key - The key (of the key-value pair) that is added in the BTreeMap field metadata of the existing Proto-Fragment’s Struct Instance

data - The hash of data is used as the value (of the key-value pair) that is added in the BTreeMap field metadata of the existing Proto-Fragment’s Struct Instance

detach

Fields

proto_hashes: Vec<[u8; 32], Global>

target_chain: SupportedChains

target_account: BoundedVec<u8, <T as Config>::DetachAccountLimit>

Request to detach a Proto-Fragment from Fragnova.

Note: The Proto-Fragment may actually get detached after one or more Fragnova blocks since when this extrinsic is called.

Note: Once the Proto-Fragment is detached, an event is emitted that includes a signature. This signature can then be used to attach the Proto-Fragment to an External Blockchain target_chain.

Arguments

origin - The origin of the extrinsic function

proto_hashes - IDs of the Proto-Fragments to detach

target_chain - External Blockchain to attach the Proto-Fragment into

target_account - Public Account Address in the External Blockchain target_chain to assign ownership of the Proto-Fragment to

ban

Fields

proto_hash: [u8; 32]

Delete Proto-Fragment proto_hash from all relevant Storage Items

Implementations

impl<T> Call<T>where
T: Config,
<T as Config>::AccountId: UncheckedFrom<<T as Config>::Hash> + AsRef<[u8]>,

pub fn new_call_variant_upload(
references: Vec<[u8; 32], Global>,
category: Categories,
tags: BoundedVec<BoundedVec<u8, <T as Config>::StringLimit>, <T as Config>::MaxTags>,
linked_asset: Option<LinkedAsset>,
license: UsageLicense<<T as Config>::AccountId>,
cluster: Option<[u8; 16]>,
data: ProtoData
) -> Call<T>

Create a call with the variant upload.

pub fn new_call_variant_patch(
proto_hash: [u8; 32],
license: Option<UsageLicense<<T as Config>::AccountId>>,
new_references: Vec<[u8; 32], Global>,
tags: Option<BoundedVec<BoundedVec<u8, <T as Config>::StringLimit>, <T as Config>::MaxTags>>,
data: Option<ProtoData>
) -> Call<T>

Create a call with the variant patch.

pub fn new_call_variant_transfer(
proto_hash: [u8; 32],
new_owner: <T as Config>::AccountId
) -> Call<T>

Create a call with the variant transfer.

pub fn new_call_variant_set_metadata(
proto_hash: [u8; 32],
metadata_key: BoundedVec<u8, <T as Config>::StringLimit>,
data: Vec<u8, Global>
) -> Call<T>

Create a call with the variant set_metadata.

pub fn new_call_variant_detach(
proto_hashes: Vec<[u8; 32], Global>,
target_chain: SupportedChains,
target_account: BoundedVec<u8, <T as Config>::DetachAccountLimit>
) -> Call<T>

Create a call with the variant detach.

pub fn new_call_variant_ban(proto_hash: [u8; 32]) -> Call<T>

Create a call with the variant ban.