Type Definition fragnova_runtime::SignedExtra

pub type SignedExtra = (CheckSpecVersion<Runtime>, CheckTxVersion<Runtime>, CheckGenesis<Runtime>, CheckEra<Runtime>, CheckNonce<Runtime>, CheckWeight<Runtime>, ChargeTransactionPayment<Runtime>);

The SignedExtension to the basic transaction logic.

What is a Signed Extension?

Substrate provides the concept of signed extensions to extend an extrinsic with additional data, provided by the SignedExtension trait.

The transaction queue regularly calls signed extensions to keep checking that a transaction is valid before it gets put in the ready queue. This is a useful safeguard for verifying that transactions won’t fail in a block. They are commonly used to enforce validation logic to protect the transaction pool from spam and replay attacks.

Source: https://docs.substrate.io/reference/transaction-format/

Footnote

Each element in the tuple implements the trait SignedExtension.

This tuple implements the trait SignedExtension. See: https://paritytech.github.io/substrate/master/sp_runtime/traits/trait.SignedExtension.html#impl-SignedExtension-for-(TupleElement0%2C%20TupleElement1%2C%20TupleElement2%2C%20TupleElement3%2C%20TupleElement4%2C%20TupleElement5%2C%20TupleElement6)

Example

Notice that in any signed transaction/extrinsic that is sent to Fragnova, it will the extra data🥕 “era”, “nonce” and “tip”: https://polkadot.js.org/apps/#/extrinsics/decode/0xf5018400d43593c715fdd31c61141abd04a99fd6822c8558854ccde39a5684e7a56da27d0150cc530a8f70343680c46687ade61e1e4cdc0dfc6d916c3143828dc588938c1934030b530d9117001260426798d380306ea3a9d04fe7b525a33053a1c31bee86750200000b000000000000003448656c6c6f2c20576f726c6421

The reason we see this additional data in the encoded extrinsic is because we have defined them here.